Sun, May 19, 2024, 5:11 PM +08:00

+++ ps9/finance/templates/temp.c
@@ -1,9 +1,9 @@
 CREATE TABLE history (
     user_id INTEGER NOT NULL,
     symbol TEXT NOT NULL,
-    price NUMERIC NOT NULL,
-    shares NUMERIC NOT NULL,
-    cash NUMERIC NOT NULL,
+    price REAL NOT NULL,
+    shares INTEGER NOT NULL,
+    cash REAL NOT NULL,
     time TIMESTAMP NOT NULL,
     FOREIGN KEY (user_id) REFERENCES users(id)
     );
